requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blis==0.7.9
brotlipy==0.7.0
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
cymem==2.0.7
dataclasses==0.8
datasets==2.7.1
debugpy==1.6.4
decorator==5.1.1
dill==0.3.6
en-core-web-trf==3.4.1
entrypoints==0.4
executing==1.2.0
filelock==3.8.2
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.14
murmurhash==1.0.9
mysql-connector-python==8.0.31
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pathy==0.10.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
platformdirs==2.5.2
preshed==3.0.8
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.2.0
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
setuptools==65.5.1
shellingham==1.5.0
six==1.16.0
smart-open==5.2.1
spacy==3.4.3
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-lookups-data==1.0.3
spacy-transformers==1.1.8
srsly==2.4.5
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
transformers==4.21.3
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.38.4
win-inet-pton==1.1.0
wrapt==1.14.1
xxhash==0.0.0
yarl==1.8.1
zipp==3.11.0
